Refactor: Split types into multiple files
